% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2roxygen.R
\name{parse_and_save}
\alias{parse_and_save}
\title{Parse the input Rd file and save the roxygen documentation into a file}
\usage{
parse_and_save(path, file, usage = FALSE)
}
\arguments{
\item{path}{the path of the Rd file}

\item{file}{the path to save the roxygen documentation}

\item{usage}{logical: whether to include the usage section in the output}
}
\value{
a character vector if \code{file} is not specified, or write the
  vector into a file
}
\description{
Parse the input Rd file and save the roxygen documentation into a file
}
\author{
Hadley Wickham; modified by Yihui Xie <\url{http://yihui.name}>
}
