\name{parse_file}
\alias{parse_file}
\title{Parse the input Rd file to a list...}
\usage{parse_file(path)}
\description{Parse the input Rd file to a list}
\details{This function uses the function \code{parse_Rd} in the \pkg{tools} package
to parse the Rd file.}
\value{a named list containing the documentation sections as strings}
\author{Hadley Wickham; modified by Yihui Xie <\url{http://yihui.name}>}
\arguments{\item{path}{the path of the Rd file}}
